+++ alerter/src/main/java/com/tapadoo/alerter/Alert.kt
@@ -1,15 +1,16 @@
 package com.tapadoo.alerter
 
 import android.annotation.TargetApi
+import android.app.Activity
 import android.content.Context
 import android.graphics.*
 import android.graphics.drawable.Drawable
 import android.os.Build
-import android.support.annotation.*
-import android.support.v4.content.ContextCompat
-import android.support.v4.view.ViewCompat
-import android.support.v7.content.res.AppCompatResources
-import android.support.v7.view.ContextThemeWrapper
+import androidx.annotation.*
+import androidx.core.content.ContextCompat
+import androidx.core.view.ViewCompat
+import androidx.appcompat.content.res.AppCompatResources
+import androidx.appcompat.view.ContextThemeWrapper
 import android.text.TextUtils
 import android.util.AttributeSet
 import android.util.Log
@@ -20,6 +21,9 @@ import android.widget.Button
 import android.widget.FrameLayout
 import android.widget.LinearLayout
 import kotlinx.android.synthetic.main.alerter_alert_view.view.*
+import android.view.DisplayCutout
+import androidx.core.view.DisplayCutoutCompat
+
 
 /**
  * Custom Alert View
@@ -65,7 +69,7 @@ class Alert @JvmOverloads constructor(context: Context, attrs: AttributeSet? = n
      * @param contentGravity Gravity of the Alert
      */
     var contentGravity: Int
-        get() = (llAlertBackground!!.layoutParams as FrameLayout.LayoutParams).gravity
+        get() = (llAlertBackground?.layoutParams as FrameLayout.LayoutParams).gravity
         set(contentGravity) {
             val paramsTitle = tvTitle?.layoutParams as LinearLayout.LayoutParams
             paramsTitle.gravity = contentGravity
@@ -100,16 +104,25 @@ class Alert @JvmOverloads constructor(context: Context, attrs: AttributeSet? = n
     }
 
     override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
-        super.onMeasure(widthMeasureSpec, heightMeasureSpec)
-
         if (!marginSet) {
             marginSet = true
 
             // Add a negative top margin to compensate for overshoot enter animation
             val params = layoutParams as ViewGroup.MarginLayoutParams
             params.topMargin = context.resources.getDimensionPixelSize(R.dimen.alerter_alert_negative_margin_top)
-            requestLayout()
+
+            // Check for Cutout
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+                val displayCutout = (context as? Activity)?.window?.decorView?.rootWindowInsets?.displayCutout
+
+                val notchHeight = displayCutout?.safeInsetTop ?: 0
+                llAlertBackground.apply {
+                    setPadding(paddingLeft, paddingTop + (notchHeight / 2), paddingRight, paddingBottom)
+                }
+            }
         }
+
+        super.onMeasure(widthMeasureSpec, heightMeasureSpec)
     }
 
     // Release resources once view is detached.