+++ app-base/src/main/java/com/tapadoo/alerter_demo/base/JavaDemoActivity.java
@@ -2,8 +2,8 @@ package com.tapadoo.alerter_demo.base;
 
 import android.graphics.Typeface;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.View;
 import android.widget.Toast;